+++ docs/CHANGELOG.md
@@ -8,10 +8,12 @@ CHANGELOG
 - Changes:
   - Use Firebase Perf to track translation and SN download and install time instead of Firebase Analytics
   - Remove Stetho (#198)
+- Refactoring:
+  - Replace broadcast channel with state flow
 - Update dependencies:
   - Gradle to 6.7
   - Android Gradle Plugin to 4.1.0
-  - Kotlin to 1.4.10
+  - Kotlin to 1.4.10, Coroutines to 1.4.0
   - AndroidX Core to 1.3.2, ConstraintLayout to 2.0.4
   - Firebase plugin to 4.3.4, Analytics to 18.0.0, Crashlytics to 17.2.2
   - Dagger Hilt to 2.29.1-alpha